+++ src/components/azure-pg/create-db.job.test.ts
@@ -69,6 +69,7 @@ test("should create a pg secret", () => {
             "restartPolicy": "Never",
           },
         },
+        "ttlSecondsAfterFinished": 86400,
       },
     }
   `